requirements.txt
alembic==1.5.2
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
flask-rebar==1.12.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==2.21.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
